// src/app/extensions/pattern-language/internet-of-things-patterns/model/internet-of-things-pattern-data.ts
import { Pattern } from 'src/app/graph/model';

export class InternetOfThingsPatternData extends Pattern {
    summary: string;
    alias: Array<string>;
    icon: string;
    problem: string;
    context: string;
    force: Array<string>;
    solution: string;
    solutionSketches: Array<string>;
    result: string;
    variants: Array<string>;
}